refactor: change ProductService into interface
