+++ src/main/java/kitchenpos/menus/tobe/application/MenuService.java
@@ -18,5 +18,7 @@ public interface MenuService {
 
 	List<Menu> findAll();
 
+	List<Menu> findAllByProductId(UUID productId);
+
 	void hideMenusBasedOnProductPrice(UUID productId);
 }
