+++ src/test/java/kitchenpos/menus/tobe/application/DefaultMenuServiceTest.java
@@ -30,6 +30,7 @@ import kitchenpos.menus.tobe.domain.MenuProduct;
 import kitchenpos.menus.tobe.domain.MenuRepository;
 import kitchenpos.menus.tobe.infra.InMemoryMenuGroupRepository;
 import kitchenpos.menus.tobe.infra.InMemoryMenuRepository;
+import kitchenpos.products.tobe.application.DefaultProductService;
 import kitchenpos.products.tobe.application.ProductService;
 import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.tobe.domain.ProductRepository;
@@ -51,7 +52,7 @@ class DefaultMenuServiceTest {
 		menuRepository = new InMemoryMenuRepository();
 		menuGroupRepository = new InMemoryMenuGroupRepository();
 		productRepository = new InMemoryProductRepository();
-		productService = new ProductService(productRepository, menuService, purgomalumClient);
+		productService = new DefaultProductService(productRepository, menuService, purgomalumClient);
 		purgomalumClient = new FakePurgomalumClient();
 		menuService = new DefaultMenuService(menuRepository, menuGroupRepository, productService, purgomalumClient);
 		menuGroupId = menuGroupRepository.save(menuGroup()).getId();
